+++ source/engine/package.json
@@ -37,6 +37,7 @@
   },
   "devDependencies": {
     "@types/minimist": "^1.2.5",
+    "@types/node": "^22.0.0",
     "@types/prompt-sync": "^4.2.3",
     "@types/source-map-support": "^0.5.10",
     "nodemon": "^3.1.2",
