--- source/engine/src/debug.js
@@ -1,6 +1,0 @@
-const debug = {};
-
-debug.GAME_PATH = 'demo_game.json';
-
-
-export default debug;
