+++ source/engine/src/debug.ts
@@ -1,0 +1,7 @@
+interface Debug {
+    GAME_PATH: string;
+}
+
+const debug: Debug = { GAME_PATH: 'demo_game.json' };
+
+export default debug;
